+++ src/actions/plip.js
@@ -165,3 +165,8 @@ export const plipsSignInfoFetched = ({ signInfo }) => ({
   type: "PLIP_PLIPS_SIGN_INFO_FETCHED",
   payload: { signInfo },
 });
+
+export const setCurrentPlip = plip => ({
+  type: "SET_CURRENT_PLIP",
+  payload: { plip },
+})