+++ src/containers/show-plip.js
@@ -10,6 +10,7 @@ import locale from "../locales/pt-BR";
 import {
   clearPlipInfo,
   fetchPlipRelatedInfo,
+  handleAppLink,
   navigate,
   navigateBack,
   openURL,
@@ -20,6 +21,7 @@ import {
 
 import {
   currentUser,
+  getCurrentPlip,
   fetchPlipRelatedInfoError,
   isFetchingPlipRelatedInfo,
   isRemainingDaysEnabled,
@@ -28,6 +30,7 @@ import {
   getPlipSignInfo,
   getUserCurrentPlipSignInfo,
   getPlipSignatureGoals,
+  handlingAppLinkError,
   hasUserJustSignedPlip,
   listRemoteConfig,
 } from "../selectors";
@@ -38,10 +41,11 @@ import PlipLayout from "../components/plip-layout";
 class Container extends Component {
   static propTypes = {
     errorFetching: PropTypes.bool,
+    errorHandlingAppLink: PropTypes.bool,
     isFetchingPlipRelatedInfo: PropTypes.bool,
     isSigning: PropTypes.bool,
     justSignedPlip: PropTypes.bool,
-    plip: PropTypes.object, // Navigation injected
+    plip: PropTypes.object,
     plipSignInfo: PropTypes.object,
     signers: PropTypes.array,
     signersTotal: PropTypes.number,
@@ -51,14 +55,17 @@ class Container extends Component {
     onOpenSigners: PropTypes.func.isRequired,
     onOpenURL: PropTypes.func.isRequired,
     onPlipSign: PropTypes.func.isRequired,
+    onRetryAppLink: PropTypes.func.isRequired,
     onShare: PropTypes.func.isRequired,
     onSignSuccessClose: PropTypes.func.isRequired,
     onViewPlip: PropTypes.func.isRequired,
   };
 
-  componentWillMount() {
+  componentWillReceiveProps(nextProps) {
     const { plip, onFetchPlipRelatedInfo } = this.props;
-    onFetchPlipRelatedInfo(plip.id);
+    if (nextProps.plip && (plip == null || plip.id !== nextProps.plip.id)) {
+      onFetchPlipRelatedInfo(nextProps.plip.id);
+    }
   }
 
   render() {
@@ -81,8 +88,9 @@ const onPlipSign = ({ dispatch, plip }) => {
   )
 };
 
-const mapStateToProps = (state, ownProps) => {
-  const plipId = ownProps.plip.id;
+const mapStateToProps = state => {
+  const currentPlip = getCurrentPlip(state);
+  const plipId = currentPlip ? currentPlip.id : null;
   const userSignInfo = getUserCurrentPlipSignInfo(state, plipId);
   let plipSignInfo = getPlipSignInfo(plipId)(state);
 
@@ -96,11 +104,13 @@ const mapStateToProps = (state, ownProps) => {
   const currentPlipShortSignersInfo = getCurrentPlipShortSignersInfo(state);
 
   return {
+    plip: getCurrentPlip(state),
     errorFetching: fetchPlipRelatedInfoError(state),
+    errorHandlingAppLink: handlingAppLinkError(state),
     isFetchingPlipRelatedInfo: isFetchingPlipRelatedInfo(state),
     isRemainingDaysEnabled: isRemainingDaysEnabled(state),
     isSigning: isSigningPlip(state),
-    justSignedPlip: hasUserJustSignedPlip(state, ownProps.plip.id),
+    justSignedPlip: plipId ? hasUserJustSignedPlip(state, plipId) : false,
     plipSignInfo: plipSignInfo,
     remoteConfig: listRemoteConfig(state),
     signers: currentPlipShortSignersInfo.users,
@@ -120,6 +130,7 @@ const mapDispatchToProps = dispatch => ({
   onOpenSigners: plipId => dispatch(navigate("signers", { plipId })),
   onOpenURL: url => dispatch(openURL(url)),
   onPlipSign: plip => onPlipSign({ dispatch, plip }),
+  onRetryAppLink: () => dispatch(handleAppLink()),
   onShare: plip => dispatch(sharePlip(plip)),
   onSignSuccessClose: plip => dispatch(removeJustSignedPlip({ plipId: plip.id })),
   onViewPlip: plip => dispatch(navigate("plipViewer", { plip })),