+++ src/services/mud-firebase-dynamic-link.android.js
@@ -6,6 +6,8 @@ export default class MUDFirebaseDynamicLink {
   subscribe(listener) {
     this.subscription = DeviceEventEmitter.addListener(onDeepLinkReceivedEvent, listener);
     NativeModules.MUDFirebaseDynamicLink.getInitialLink();
+
+    return () => this.unsubscribe();
   }
 
   unsubscribe() {